Update Pull Request guidlines in Traefik/ traefik
We have three sources of truth for our contributing docs.  This updates the source in Traefik/traefik to match the source in doc.traefik.io

I have simply copied and pasted the markdown from the approved source to here.  This does include the title and description information which you may want to adjust.